Move additional charges and return link into base
